fix: cast matrix variables to Float32Array in WebGL samples
Updated multiple WebGL sample files to cast matrix variables to Float32Array for improved type safety and compatibility. This change ensures that the binding resources are correctly typed, enhancing the stability of the rendering process.

+++ examples/src/WebGL2Samples/fbo_multisample.ts
@@ -107,7 +107,7 @@ const renderPass1: RenderPass = {
             ...programs[PROGRAM.TEXTURE],
             primitive: { topology: "LINE_LOOP" },
         },
-        bindingResources: { MVP: IDENTITY },
+        bindingResources: { MVP: IDENTITY as Float32Array },
         vertices: vertexArrays[PROGRAM.TEXTURE].vertices,
         draw: { __type__: "DrawVertex", vertexCount },
     }]
@@ -125,7 +125,7 @@ const renderPass2: RenderPass = {
     renderPassObjects: [
         {
             pipeline: programs[PROGRAM.SPLASH],
-            bindingResources: { diffuse: { texture, sampler }, MVP: mvp },
+            bindingResources: { diffuse: { texture, sampler }, MVP: mvp as Float32Array },
             vertices: vertexArrays[PROGRAM.SPLASH].vertices,
             draw: { __type__: "DrawVertex", vertexCount: 6 },
         }
