+++ examples/src/WebGL2Samples/texture_grad.ts
@@ -221,8 +221,8 @@ import { getShaderSource, loadImage } from "./utility";
         mat4.rotateY(mvMatrix, mvMatrix, orientation[1] * Math.PI);
         mat4.rotateZ(mvMatrix, mvMatrix, orientation[2] * Math.PI);
 
-        reactive(ro.bindingResources).mvMatrix = mvMatrix;
-        reactive(ro.bindingResources).pMatrix = perspectiveMatrix;
+        reactive(ro.bindingResources).mvMatrix = mvMatrix as Float32Array;
+        reactive(ro.bindingResources).pMatrix = perspectiveMatrix as Float32Array;
         reactive(ro.bindingResources).diffuse = { texture, sampler };
 
         webgl.submit({ commandEncoders: [{ passEncoders: [rp] }] });
