+++ examples/src/WebGL2Samples/texture_vertex.ts
@@ -216,8 +216,8 @@ import { getShaderSource, loadImage } from "./utility";
                         primitive: { topology: IDrawMode2Name[primitive.mode] },
                     },
                     bindingResources: {
-                        mvMatrix: localMV,
-                        pMatrix: perspectiveMatrix,
+                        mvMatrix: localMV as Float32Array,
+                        pMatrix: perspectiveMatrix as Float32Array,
                         displacementMap: { texture, sampler },
                         diffuse: { texture, sampler },
                     },
